scheduling update of share price works correctly

+++ src/main/java/fp/stock/share/Share.java
@@ -16,5 +16,5 @@ public class Share {
     private Long id;
     private String name;
     private Long quantity;
-    private Long unitPrice;
+    private Double unitPrice;
 }
